Update uploadFileContent for Python client
This updates uploadFileContent to handle the needs of both the Go and
Python clients when uploading over multiple requests. It rewrites
parseRange as parseContentRange, to parse the header content and leave
interpretation and setting response headers to uploadFileContent.

The Go client always sends content as one or more chunks without a total
size, and, when the stream is exhausted, sends a final request with no
content but the final size.

The Go client also sends the header X-Guploader-No-308. This specifies
that the response to a "mid-stream" request should be sent with a
status code 200 OK plus the header X-Http-Status-Code-Override 308.

The Python client client sends chunks with a total size when known, such
as when uploading a file or an in-memory string.  Unlike a streaming
response, the final request contains content. It expects a mid-stream
response to have a status code 308 Permanent Redirect.

The Python client is pickier about some response details, and
uploadFileContent's responses are changed to work with both the Python
and (more flexible) Go clients:

 * The Python client requires the "Range" response header to start
   counting at 0, such as "bytes 0-999", while the Go client appears to
   ignore the start byte.
 * The Python client requires that the final response is a "200 OK", while
   the Go client accepts this or a "201 Created".