Disable gzipped responses
The Java SDK sends gzipped requests, and some clients will send the
appropriate accept headers, but we don't _need_ to send gzipped
responses. This makes the server simpler and easier to debug. Plus,
supporting the Java SDK doesn't require gzipped responses.

The idea here is to debug and potentially use this as the fix for #419.
This should have no effect on gzipped _requests_.