Improve memory usage and execution time of listing objects with file system backend (#556)
* Do not store all objects in memory on list commands

Previously all of the object blobs were read into memory on every list
command. Even when the list command would've returned nothing. There are
many problems with this approach:

* It is extremely slow to read all files on every list
* At some point if the blobs in the storage are more than the memory on
the machine the process will crash with OOM error.

This MR makes it so that for most operations the actual blob object
contents are not kept in memory. Instead only a small struct
(ObjectAttrs) is used. Unfortunately due to the nature of JSON encoding
all objects are read at least once from the disk in full.

* List objects: filter files by prefix before reading them

Previously all files were read from a bucket before some of them were
dropped after prefix test. This is extremely inefficient in light of the
fact that all files are actually read into memory for the file system
bucket. A moderatly large bucket will cause listing to take many minutes
even when eventually only a few results are returned.