Store object metadata and data separately
The former remains a JSON-encoded blob while the latter is stored in its
native format.  This reduces memory usage and is much faster by avoiding
JSON-encoding large objects.  This enables a future commit to avoid
reading the entire object, particularly for range requests.  Note that
this commit changes the on-disk format and is not compatible with
previous data sets.  References #669.  Fixes #671.