Add HEAD support for storage API object endpoint (#759)
* Support HEAD method on storage API

Problem: As reporeted in #667. The storage API is replying 404 for
existing objects when method HEAD is used, although GET replies with
200.

Solution: Update storage API method route and handler to behave as
"donwload" when HEAD method is used.

* Update documentation regarding flag usage

Problem: As a new user, one tends to simply use what comes in the
README.md file without realizing there is plenty more features at hand.

Solution: Update README.md to pointing out there is a -help flag that
lists all the available configuration overrides. I was tempted to copy
and paste its output, but it might add too much noise for the reader
so disregarded that thought.