Propagate ContentType field on create/modify ops
This commit tries to address 2 bugs that where found when using
some of the create/modify operations GCS supports like:
* using the streamable upload operations of GCS to upload of a File.
* update the blob metadata content-type without modifying the bytes of
  the blob.

For the streamable upload, by default if no content-type is provided;
GCS sets it up to 'application/octet-stream' or uses the one provided by
the user in the BlobInfo (This is the expected behaviour for this operation in GCS).

For the update, when we merge metadata information regarding the
content-type, it should override it the update response should propagate
this change.