Unify logic for seeding the server from a directory
This should fix the flaky test: both the main and the fakestorage
packages would create that directory and delete it, and since package
tests run in parallel, there was a race between deleting and running the
tests for generateObjectsFromFiles. I'm surprise we only observe
flakiness on Windows.

On top of fixing the test, this also reduces code duplication. One
downside is that we lose an info log, but I'm not too moved by that.