Remove unnecessary quoting around ETag in JSON response
The JSON marshaller will already escape the value appropriately. Quoting
the value explicitly beforehand is counterproductive, as the quotes
introduced in that step would be considered legitimate part of the ETag
value and would be themselves escaped when marshalling.

When setting the `ETag` HTTP response header, however, we do have to
explicitly add the quoting. This is because double quotes are not added
automatically for HTTP headers and, according to the [standard][1], the
`ETag` header value _must_ be wrapped in double quotes.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/ETag