Additional object response attributes
Some GCS client libraries (like [this one][1]), are quite strict when it
comes to the set of attributes they expect to find in the response. If
any of those attributes is missing, they fail to deserialize the
response.

Here we introduce three new attributes in the Object response:
* `SelfLink`
* `MediaLink`
* `Metageneration`

`SelfLink` cannot be tested using the Go GCS client library, as it is
not included in the [`ObjectAttrs`][2] struct.

For the `Metageneration` we just hardcode the value `"1"`, considering
that we are not keeping track of it internally and `"1"` happens to be
the correct value most of the times anyway.

[1]: https://crates.io/crates/google-cloud-storage
[2]: https://pkg.go.dev/cloud.google.com/go/storage@v1.39.0#ObjectAttrs