Don't omit `versioning.enabled` in bucket response
In Go's `json` package, `false` qualifies as _empty_ value for the
purposes of marshalling (see `Marshal` [docs][1]).

In the bucket JSON response, the `versioning.enabled` attribute should
still be present even when it is `false`. If omitted, some [client
libraries][2] fail to deserialize the response.

This commit changes the marshalling of `versioning` from:

```json
"versioning": {}
```

to:

```json
"versioning": {
  "enabled": false
}
```

[1]: https://pkg.go.dev/encoding/json#Marshal
[2]: https://crates.io/crates/google-cloud-storage