Additional ACL response attributes
Include these attributes in the JSON response for access control lists.

* `etag`
* `kind`

Some [client libraries][1] expect them to be present, and fail to
deserialize the response otherwise.

The value for `kind` matches what the real GCS API returns. We don't
care too much about the ETag value at this stage, so we just picked a
valid constant.

[1]: https://crates.io/crates/google-cloud-storage